Extract propsMapper()
